requirements.txt
Flask==3.0.0
google-generativeai==0.3.2
python-dotenv==1.0.0
pytest==7.4.3
pytest-mock==3.12.0
gunicorn==21.2.0